+++ Source/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(Editors)
 add_subdirectory(Engine)
 add_subdirectory(Game)
+add_subdirectory(Render)
 add_subdirectory(Utils)
