+++ Source/Engine/CMakeLists.txt
@@ -11,14 +11,14 @@ target_include_directories(${PROJECT_NAME}
         "${CMAKE_CURRENT_SOURCE_DIR}/include"
 )
 
-find_package(GLEW CONFIG REQUIRED)
+#find_package(GLEW CONFIG REQUIRED)
 find_package(glfw3 CONFIG REQUIRED)
 find_package(EnTT CONFIG REQUIRED)
 find_package(Bullet CONFIG REQUIRED)
 find_package(imgui CONFIG REQUIRED)
 find_package(imguizmo CONFIG REQUIRED)
-find_package(glm CONFIG REQUIRED)
-find_package(SailC++ CONFIG REQUIRED)
+#find_package(glm CONFIG REQUIRED)
+#find_package(SailC++ CONFIG REQUIRED)
 find_package(spdlog CONFIG REQUIRED)
 
 target_link_libraries(${PROJECT_NAME}
@@ -27,14 +27,20 @@ target_link_libraries(${PROJECT_NAME}
 target_include_directories(${PROJECT_NAME} PRIVATE {CMAKE_SOURCE_DIR}/Source/Utils/ModelLoader/include)
 
 target_link_libraries(${PROJECT_NAME}
+        LINK_PUBLIC
+        Render
+)
+target_include_directories(${PROJECT_NAME} PRIVATE {CMAKE_SOURCE_DIR}/Source/Render/include)
+
+target_link_libraries(${PROJECT_NAME}
         PUBLIC
-        GLEW::GLEW
+#        GLEW::GLEW
         glfw
         EnTT::EnTT
         LinearMath Bullet3Common BulletDynamics BulletSoftBody
         imgui::imgui
         imguizmo::imguizmo
-        glm::glm
-        SAIL::sail-c++
+#        glm::glm
+#        SAIL::sail-c++
         spdlog::spdlog spdlog::spdlog_header_only
 )
