--- Source/Engine/include/Graphics/GLTypeSize.h
@@ -1,14 +1,0 @@
-//
-// Created by dryerem19 on 10/25/22.
-//
-
-#pragma once
-#include "core.h"
-
-namespace KitEngine::Graphics
-{
-    enum GLTypeSize
-    {
-        Float = sizeof(GLfloat),
-    };
-}
