--- Source/Engine/include/Graphics/VertexBufferAttribute.h
@@ -1,18 +1,0 @@
-//
-// Created by dryerem19 on 10/24/22.
-//
-
-#pragma once
-#include "core.h"
-#include "GLTypeSize.h"
-
-namespace KitEngine::Graphics
-{
-    struct VertexBufferAttribute
-    {
-        GLsizei      Count;
-        GLsizei      Type;
-        GLboolean    Normalized;
-        GLTypeSize   TypeSize;
-    };
-}
