--- Source/Engine/source/Graphics/Renderer.cpp
@@ -1,62 +1,0 @@
-//
-// Created by dryerem19 on 21.10.22.
-//
-#include "Graphics/Renderer.h"
-#include "Graphics/GlHelpers.h"
-
-KitEngine::Graphics::Renderer &KitEngine::Graphics::Renderer::Instance() {
-
-    static Renderer instance;
-    return instance;
-
-}
-
-void KitEngine::Graphics::Renderer::Clear() {
-
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-}
-
-void KitEngine::Graphics::Renderer::Draw(const KitEngine::Graphics::VertexArray &vertexArray,
-                                         const KitEngine::Graphics::IndexBuffer &indexBuffer,
-                                         const KitEngine::Graphics::Shader &shader) const {
-
-    vertexArray.Bind();
-    //indexBuffer.Bind();
-    GLCall(glDrawElements(GL_TRIANGLES, indexBuffer.GetCount(),
-                          GL_UNSIGNED_INT, nullptr));
-
-}
-
-void KitEngine::Graphics::Renderer::DrawModel(const KitEngine::Graphics::Components::ModelComponent &model,
-                                              const KitEngine::Graphics::Shader &shader,
-                                              const std::vector<KitEngine::Graphics::Texture> &textures) {
-
-    int i = 0;
-
-
-    for (auto& mesh : model.mMeshes) {
-
-        model.mVertexArray.Bind();
-        textures[mesh.MaterialIndex].Enable();
-
-        GLCall(glDrawElementsBaseVertex(GL_TRIANGLES,
-                                        mesh.NumIndices,
-                                        GL_UNSIGNED_INT,
-                                        reinterpret_cast<const void *>(sizeof(unsigned int) *
-                                                                       mesh.BaseIndex),
-                                        mesh.BaseVertex));
-
-        model.mVertexArray.Unbind();
-        textures[mesh.MaterialIndex].Disable();
-    }
-    KitEngine::Graphics::VertexArray::Unbind();
-
-}
-
-void KitEngine::Graphics::Renderer::Initialize() {
-
-    glEnable(GL_DEPTH_TEST);
-
-}
-
