+++ Source/Render/pch/RenderPch.h
@@ -1,0 +1,23 @@
+//
+// Created by dryerem19 on 08.11.22.
+//
+
+#ifndef KITENGINE_RENDERPCH_H
+#define KITENGINE_RENDERPCH_H
+
+#include <iostream>
+#include <sstream>
+#include <fstream>
+#include <cstdint>
+#include <string>
+
+#include <sail/sail-c++/sail-c++.h>
+#include <GL/glew.h>
+
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtc/type_ptr.hpp>
+
+#include "Helpers.h"
+
+#endif //KITENGINE_RENDERPCH_H
