+++ bin/Debug/imgui.ini
@@ -1,0 +1,10 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Window][Window]
+Pos=552,437
+Size=334,280
+Collapsed=0
+
